Implement 0.1.3 sandbox

+++ app/src/main/java/paymaya/com/paymayaandroidcheckout/utils/Utils.java
@@ -3,13 +3,10 @@ package paymaya.com.paymayaandroidcheckout.utils;
 import android.content.Context;
 import android.widget.ImageView;
 
+import com.paymaya.checkoutsdkandroid.models.Item;
+import com.paymaya.checkoutsdkandroid.models.TotalAmount;
 import com.squareup.picasso.MemoryPolicy;
 import com.squareup.picasso.Picasso;
-import com.voyagerinnovation.paymaya_sdk_android_checkout.models.Address;
-import com.voyagerinnovation.paymaya_sdk_android_checkout.models.Buyer;
-import com.voyagerinnovation.paymaya_sdk_android_checkout.models.Contact;
-import com.voyagerinnovation.paymaya_sdk_android_checkout.models.Item;
-import com.voyagerinnovation.paymaya_sdk_android_checkout.models.TotalAmount;
 
 import java.io.File;
 import java.math.BigDecimal;
@@ -32,7 +29,7 @@ public class Utils {
         //item1
         ItemModel itemModel = new ItemModel();
         TotalAmount totalAmount = new TotalAmount(BigDecimal.valueOf(803.78), "PHP");
-        Item item = new Item("Bag One", BigDecimal.valueOf(1), totalAmount);
+        Item item = new Item("Bag One", 1, totalAmount);
         item.setSkuCode("CVG-096731");
         item.setDescription("bag");
 
@@ -43,7 +40,7 @@ public class Utils {
         //item2
         itemModel = new ItemModel();
         totalAmount = new TotalAmount(BigDecimal.valueOf(1945.45), "PHP");
-        item = new Item("Shoe One", BigDecimal.valueOf(1), totalAmount);
+        item = new Item("Shoe One", 1, totalAmount);
         item.setSkuCode("CVG-096732");
         item.setDescription("shoe");
 
@@ -54,7 +51,7 @@ public class Utils {
         //item3
         itemModel = new ItemModel();
         totalAmount = new TotalAmount(BigDecimal.valueOf(5892.34), "PHP");
-        item = new Item("Necklace One", BigDecimal.valueOf(1), totalAmount);
+        item = new Item("Necklace One", 1, totalAmount);
         item.setSkuCode("CVG-096733");
         item.setDescription("necklace");
 
@@ -65,7 +62,7 @@ public class Utils {
         //item4
         itemModel = new ItemModel();
         totalAmount = new TotalAmount(BigDecimal.valueOf(545.56), "PHP");
-        item = new Item("Tshirt One", BigDecimal.valueOf(1), totalAmount);
+        item = new Item("Tshirt One", 1, totalAmount);
         item.setSkuCode("CVG-096734");
         item.setDescription("tshirt");
 
@@ -76,7 +73,7 @@ public class Utils {
         //item5
         itemModel = new ItemModel();
         totalAmount = new TotalAmount(BigDecimal.valueOf(645.67), "PHP");
-        item = new Item("Tshirt Two", BigDecimal.valueOf(1), totalAmount);
+        item = new Item("Tshirt Two", 1, totalAmount);
         item.setSkuCode("CVG-096735");
         item.setDescription("tshirt");
 
@@ -87,7 +84,7 @@ public class Utils {
         //item6
         itemModel = new ItemModel();
         totalAmount = new TotalAmount(BigDecimal.valueOf(1645.03), "PHP");
-        item = new Item("Shoe Two", BigDecimal.valueOf(1), totalAmount);
+        item = new Item("Shoe Two", 1, totalAmount);
         item.setSkuCode("CVG-096736");
         item.setDescription("shoe");
